Restores editability of perspective widget from perspective python when
running in JupyterLab.

This commit removes the `editable` attribute from PerspectiveWidget and
PerspectiveTraitlets, bringing it inline with the changes made to the
PerspectiveViewer. The Perspective JupyterLab extension is reimplemented
to leverage `plugin_config` for editability instead of `editable`, and
tests are greatly extended to test the various js<->python
synchronizeable attributes.