Always install current version in install_emsdk.js
The build error you get when emsdk has drifted out of date looks like:

    error: error: tool is not installed and therefore cannot be activated: 'releases-adedc0750c4a89b65bee866edab24298cb8d6677-64bit'
    Command failed: cd /Users/tom/perspective/perspective/.emsdk && . ./emsdk_env.sh >/dev/null 2>&1 && emsdk activate 3.1.36 >/dev/null && cd /Users/tom/perspective/perspective/cpp/perspective && node ./build.js

The resolution to this now is to run `node scripts/install_emsdk.js`.
Previously it wouldn't install the current emsdk version unless the
developer first rm-rfed .emsdk

It's also still cheap to run the script when the current version is
already installed.