replace find_package(Boost) with psp_build_dep
This may make local builds from clean take slightly longer (since it
won't use any latent Boost on the system), but should make CI builds
faster since the "build boost" step of install-tools.mjs can be skipped.
The headers in Boost's release tarballs can be used directly.

Was also possible to update `find_package(Boost)` call to use `PATHS` to
request the Emscripten build to look in system trees, which I tried
first, but getting the paths to where FindBoost.cmake might be right on
every environment was like whack-a-mole.  This way is the same on every
environment and makes Boost dependency consistent with the others