arrow: Fix cross-compiling errors on macos
This fixes the issue with missing LZ4 symbols in the arm64
wheel first introduced in v3.1.1.

The crux of the issue is that Arrow was building lz4 (and zstd) in a
separate cmake build, and it does not specially pass on the
`CMAKE_OSX_ARCHITECTURES` flag that was set when perspective-server's
build script invokes cmake.

Arrow's build does forward on the value of CMAKE_TOOLCHAIN_FILE to its
external dependencies, if it's set, so this commit creates two minimal
toolchain files and uses them in the Mac build to set
`CMAKE_OSX_ARCHITECTURES` (and a couple other variables) correctly when
cross-compiling.

a good follow-up would be to see if we can enable Python tests on
arm64 using a macos-14 runner, which is available for public repos:
<https://docs.github.com/en/actions/using-github-hosted-runners/using-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories>

Signed-off-by: Tom Jakubowski <tom@prospective.dev>