sdist: include data_dir again, license dir fix
The data dir was accidentally removed from the sdist in 053abcd.

Our Maturin version is currently incorrectly packaging the license files
under `.dist-info/license_files`, when it should be (according to PEP
639) in `.dist-info/licenses`.  This has been fixed in newer versions of
Maturin.  Update the CI test to check in both places so that it won't
break when we update Maturin.

Signed-off-by: Tom Jakubowski <tom@prospective.dev>