Fix date filters in perspective-{select,click} events
Fixes issue #2949 where date filters in the `perspective-select` and
`perspective-click` events use a zero-indexed month, but the date filter
input uses a one-indexed month. The date filter returned by
`perspective-config-update` is correctly one-indexed. This issue makes
the filtered data in a workspace incorrect. It also makes perspective
crash if a January 1 date gets converted to `2025-00-01` and then passed
to a second view.

Fix this by simply adding 1 to the month when generating the config in
`ProtoServer::_handle_request`.

Signed-off-by: Nicholas Burlett <nburlett@ghs.com>