Add test for perspective-click date bug
Regression test for issue #2949 where date filters are not correctly
formatted on the perspective-click and perspective-select custom events.
The test only exercises perspective-click for simplicity.

A new all-types-small-multi-arrow-test.html was necessary, as the issue
only affects the date datatype, and the existing csv tests only exercise
the datetime data type.

Signed-off-by: Nicholas Burlett <nburlett@ghs.com>